Add files via upload
This dataset shows the answers of 10 respondents regarding the increasing mental health epidemics in the world. These respondents were asked their opinions on the leading causes of mental health issues (and why they cause them), their favorite and least favorite aspects on modern human life, and the factors that harm the mental health of mothers, fathers, and children.